requirements.txt
pip
autopep8
fastapi
gunicorn==22.0.0
Werkzeug==3.0.6
uvicorn
python-dotenv
google-cloud-storage
google-cloud-aiplatform
google-generativeai
pymupdf
python-multipart
google-cloud-translate==3.13.0
langchain==0.2.1
langchain-text-splitters==0.2.0